Rename module from graph to display
