+++ trustchain-core/src/display.rs
